react-redux defalt setting
