+++ client/src/_reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from "redux";
+//import user from "./user_reducer"
+
+const rootReducer = combineReducers({
+    //user,
+});
+
+export default rootReducer;
