+++ client/src/components/views/LandingPage/LandingPage.js
@@ -1,4 +1,5 @@
-import React from "react";
+import React, { useEffect } from "react";
+import axios from "axios";
 
 function LandingPage() {
     return <div>hi</div>;
