+++ client/src/components/views/SignupPage/SIgnupPage.js
@@ -1,0 +1,20 @@
+//import React, { useEffect } from "react";
+//import axisos from "axios";
+import { Button } from "antd";
+
+function SIgnupPage() {
+    //useEffect(() => {
+    //    axisos.get("/users/signup").then((res) => console.log(res));
+    //}, []);
+    return (
+        <div>
+            <form>
+                <input type="text"></input>
+                <input type="text"></input>
+                <Button type="primary">Button</Button>
+            </form>
+        </div>
+    );
+}
+
+export default SIgnupPage;
