--- front/src/components/views/SignupPage/SIgnupPage.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function SIgnupPage() {
-    return <div></div>;
-}
-
-export default SIgnupPage;
